services/bot/requirements.txt
aiogram==3.6.0
httpx==0.27.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.7
apscheduler==3.10.4
alembic==1.13.2
sqlalchemy==2.0.30
